Switch to the 2024 edition of Rust
Before this change, if you ran “cargo check”, then you would get this
warning:

> warning: no edition set: defaulting to the 2015 edition while the
> latest is 2024

This change fixes that warning. The main motivation behind this change
is to prepare for a future commit. That future commit will change this
project so that it contains more Rust code. (At the moment, this project
only contains two lines of Rust code). It makes more sense to write new
Rust code using the latest edition of Rust than it does to write new
Rust code using the oldest edition of Rust.